getMember util
