+++ commands/economy/inventory.js
@@ -1,5 +1,4 @@
 const Items = require('../../models/userItems');
-const { getMember } = require('../utils/getMember');
 
 module.exports = {
 	config: {
@@ -10,7 +9,7 @@ module.exports = {
 		aliases: ['inv'],
 		channel: 'bot-commands'
 	},
-	run: async ({ client: bot, message, args }) => {
+	run: async ({ client: bot, message, args, utils: { getMember } }) => {
 
 		const user = getMember(message, args.join(' '));
