+++ commands/fun/skillrate.js
@@ -1,4 +1,3 @@
-const { getMember } = require('../utils/getMember');
 const userData = require('../../models/userData.js');
 const userQuests = require('../../models/userQuests');
 
@@ -11,7 +10,7 @@ module.exports = {
 		aliases: ['prorate', 'skill', 'pro', 'rate'],
 		channel: 'bot-commands'
 	},
-	run: async ({ client: bot, message, args }) => {
+	run: async ({ client: bot, message, args, utils: { getMember } }) => {
 
 		const member = getMember(message, args.join(' '));
 		const rating = Math.floor(Math.random() * 101);
